rename MainPage.py to main_page.py
(and update all calls to it)
