+++ hint tool.py
@@ -6,7 +6,7 @@ import sys
 # Don't write __pycache__
 sys.dont_write_bytecode = True
 
-from hints.gui.MainPage import create_default_notebook, create_pop_up_buttons
+from hints.gui.main_page import create_default_notebook, create_pop_up_buttons
 from hints.gui.Utils import create_notebook_tab
 from hints.gui.shopping.agitha import AgithaTab
 from hints.gui.shopping.jovani import JovaniTab
